feat: refactor code and serialized account output
